Relocate DiskSpace metrics to SystemMetricsAutoConfiguration
Closes gh-27688
