+++ spring-boot-project/spring-boot-actuator-autoconfigure/src/main/java/org/springframework/boot/actuate/autoconfigure/metrics/SystemMetricsAutoConfiguration.java
@@ -16,7 +16,10 @@
 
 package org.springframework.boot.actuate.autoconfigure.metrics;
 
+import java.io.File;
+
 import io.micrometer.core.instrument.MeterRegistry;
+import io.micrometer.core.instrument.binder.jvm.DiskSpaceMetrics;
 import io.micrometer.core.instrument.binder.system.FileDescriptorMetrics;
 import io.micrometer.core.instrument.binder.system.ProcessorMetrics;
 import io.micrometer.core.instrument.binder.system.UptimeMetrics;
@@ -57,6 +60,12 @@ public class SystemMetricsAutoConfiguration {
 	@ConditionalOnMissingBean
 	public FileDescriptorMetrics fileDescriptorMetrics() {
 		return new FileDescriptorMetrics();
+	}
+
+	@Bean
+	@ConditionalOnMissingBean
+	public DiskSpaceMetrics diskSpaceMetrics() {
+		return new DiskSpaceMetrics(new File("."));
 	}
 
 }
